TEST: Move ucs/gtest to test/gtest
