+++ test/gtest/ucm/malloc_hook.cc
@@ -10,8 +10,8 @@
 
 #include <ucs/arch/atomic.h>
 #include <ucs/type/status.h>
-#include <ucs/gtest/test.h>
-#include <ucs/gtest/test_helpers.h>
+#include <common/test.h>
+#include <common/test_helpers.h>
 #include <pthread.h>
 #include <sstream>
 #include <stdint.h>
