+++ test/gtest/ucp/test_ucp_tag.cc
@@ -7,7 +7,7 @@
 
 #include "ucp_test.h"
 
-#include <ucs/gtest/test_helpers.h>
+#include <common/test_helpers.h>
 extern "C" {
 #include <ucs/debug/debug.h>
 }
