+++ test/gtest/ucs/test_arbiter.cc
@@ -5,8 +5,8 @@
 */
 
 
-#include <ucs/gtest/test.h>
-#include <ucs/gtest/test_helpers.h>
+#include <common/test.h>
+#include <common/test_helpers.h>
 
 extern "C" {
 #include <ucs/sys/sys.h>
