+++ test/gtest/ucs/test_async.cc
@@ -4,8 +4,8 @@
  * See file LICENSE for terms.
  */
 
-#include <ucs/gtest/test.h>
-#include <ucs/gtest/test_helpers.h>
+#include <common/test.h>
+#include <common/test_helpers.h>
 
 extern "C" {
 #include <ucs/async/async.h>
