+++ test/gtest/ucs/test_config.cc
@@ -4,7 +4,7 @@
 * See file LICENSE for terms.
 */
 
-#include <ucs/gtest/test.h>
+#include <common/test.h>
 extern "C" {
 #include <ucs/config/parser.h>
 #include <ucs/time/time.h>
