+++ test/gtest/ucs/test_datatype.cc
@@ -4,7 +4,7 @@
 * See file LICENSE for terms.
 */
 
-#include <ucs/gtest/test.h>
+#include <common/test.h>
 extern "C" {
 #include <ucs/datastruct/notifier.h>
 #include <ucs/datastruct/list.h>
