+++ test/gtest/ucs/test_frag_list.cc
@@ -4,7 +4,7 @@
 * See file LICENSE for terms.
 */
 
-#include <ucs/gtest/test.h>
+#include <common/test.h>
 extern "C" {
 #include <ucs/datastruct/frag_list.h>
 }
